move temporary tags to conftest
